+++ resources/views/note/edit.blade.php
@@ -3,7 +3,7 @@
 @section('content')
     <form id="note-edit" action="/note-update/{{ $context->model->id }}" method="POST">
         @csrf
-        <x-laraknife.panels.standard title="{{ __('Change of a Note') }}">
+        <x-laraknife.panels.edit title="{{ __('Change of a Note') }}">
             <x-laraknife.layout.nav-tabs :info="$navTabsInfo">
                 <x-laraknife.forms.combobox position="first" name="category_scope" label="Category" :options="$optionsCategory"
                     value="{{ $context->valueOf('category_scope') }}" width2="4" />
@@ -18,6 +18,6 @@
                     <x-laraknife.buttons.cancel width1="2" width2="4" />
                 </div>
             </x-laraknife.layout.nav-tabs>
-        </x-laraknife.panels.standard>
+        </x-laraknife.panels.edit>
     </form>
 @endsection