+++ resources/views/user/login.blade.php
@@ -14,6 +14,8 @@
             <x-laraknife.forms.set-position position="first" />
             <x-laraknife.layout.col-empty width="3" />
             <x-laraknife.buttons.button-position position="last" name="btnLogin" label="{{ __('Login') }}" />
+            <x-laraknife.layout.row-empty />
+            <x-laraknife.forms.link position="alone" reference="/user-forgotten" text="{{__('Password forgotten?')}}" width1="5" width2="4" />
         </x-laraknife.panels.standard>
     </form>
 @endsection